Move fasttrack navigation action factory into its own service for clarity. TODOs remaining.
