+++ app/js/app/controllers/QuestionPageControllers.js
@@ -105,6 +105,7 @@ define([], function() {
 		persistence.session.save("conceptPageSource", $location.url());
 
 		if ($stateParams.board) {
+			$scope.gameboardId = $stateParams.board;
 			$scope.gameBoard = api.gameBoards.get({id: $stateParams.board});
 			$scope.gameBoard.$promise.then(function(board) {
 
@@ -136,7 +137,7 @@ define([], function() {
 		}
 
 		$scope.questionHistory = $stateParams.questionHistory ? $stateParams.questionHistory.split(',') : [];
-
+		
 		$scope.backToBoard = function() {
 			$location.url("/gameboards#" + $stateParams.board)
 		}
